moved styles to another file and mate folder for every component because of scaling
