+++ src/Components/Navbar/ToggleTheme/ToggleStyles.tsx
@@ -1,0 +1,27 @@
+import { createStyles } from "@mantine/core";
+
+export const useStyles = createStyles((theme) => ({
+    root: {
+        position: "relative",
+        "& *": {
+            cursor: "pointer",
+        },
+    },
+
+    icon: {
+        pointerEvents: "none",
+        position: "absolute",
+        zIndex: 1,
+        top: 3,
+    },
+
+    iconLight: {
+        left: 4,
+        color: theme.white,
+    },
+
+    iconDark: {
+        right: 4,
+        color: theme.colors.gray[6],
+    },
+}));
