+++ src/Components/Navbar/ToggleTheme/ToggleTheme.tsx
@@ -1,36 +1,6 @@
-import {
-    createStyles,
-    Switch,
-    Group,
-    useMantineColorScheme,
-} from "@mantine/core";
+import { Switch, Group, useMantineColorScheme } from "@mantine/core";
 import { IconSun, IconMoonStars } from "@tabler/icons";
-
-const useStyles = createStyles((theme) => ({
-    root: {
-        position: "relative",
-        "& *": {
-            cursor: "pointer",
-        },
-    },
-
-    icon: {
-        pointerEvents: "none",
-        position: "absolute",
-        zIndex: 1,
-        top: 3,
-    },
-
-    iconLight: {
-        left: 4,
-        color: theme.white,
-    },
-
-    iconDark: {
-        right: 4,
-        color: theme.colors.gray[6],
-    },
-}));
+import { useStyles } from "./ToggleStyles";
 
 export function SwitchToggle() {
     const { colorScheme, toggleColorScheme } = useMantineColorScheme();
