+++ src/Components/News/FilteringComponents/LoadMore/LoadMoreStyles.tsx
@@ -1,0 +1,9 @@
+import { createStyles } from "@mantine/core";
+
+export const useStyles = createStyles(() => ({
+    buttonwrapper: {
+        width: "100%",
+        textAlign: "center",
+        marginBottom: "3%",
+    },
+}));
