+++ src/Components/News/FilteringComponents/Mapping/MappingStyles.tsx
@@ -1,0 +1,17 @@
+import { createStyles } from "@mantine/core";
+
+export const useStyles = createStyles(() => ({
+    container: {
+        width: "100%",
+        display: "flex",
+        padding: "5% 0%",
+        flexWrap: "wrap",
+        justifyContent: "center",
+        alignItems: "center",
+    },
+    buttonwrapper: {
+        width: "100%",
+        textAlign: "center",
+        marginBottom: "3%",
+    },
+}));
