+++ src/Components/News/FullPageSub/FullPageContent/FullPageContent.tsx
@@ -1,36 +1,5 @@
 import { Text, Title, Image } from "@mantine/core";
-import { createStyles } from "@mantine/core";
-
-const useStyles = createStyles((theme, _params, getRef) => ({
-    fullpage: {
-        width: "100%",
-        marginTop: "3%",
-        display: "flex",
-        flexDirection: "column",
-        alignItems: "center",
-        padding: "0% 15%",
-    },
-    imagewrapper: {
-        width: "60%",
-        margin: "3% 0%",
-    },
-    title: {
-        textAlign: "center",
-    },
-    date: {
-        margin: "1% 0%",
-    },
-    description: {
-        fontSize: "20px",
-    },
-    sources: {
-        marginTop: "5%",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "space-evenly",
-        width: "100%",
-    },
-}));
+import { useStyles } from "./PageContentStyles";
 export const FullPageContent = (props: { data: any }) => {
     const { classes } = useStyles();
     return (
