+++ src/Components/News/SearchSub/Button/Button.tsx
@@ -1,22 +1,5 @@
-import { createStyles, Button } from "@mantine/core";
-
-const useStyles = createStyles(() => ({
-    input: {
-        width: "50%",
-    },
-    buttonWrapper: {
-        width: "100%",
-        display: "flex",
-        justifyContent: "center",
-        marginTop: "3%",
-        padding: "0% 43%",
-        height: "40px",
-        color: "primary",
-    },
-    button: {
-        width: "100%",
-    },
-}));
+import { Button } from "@mantine/core";
+import { useStyles } from "./ButtonStyles";
 
 interface SearchButtonType {
     term: string;
