+++ src/Components/News/SearchSub/Button/ButtonStyles.tsx
@@ -1,0 +1,19 @@
+import { createStyles } from "@mantine/core";
+
+export const useStyles = createStyles(() => ({
+    input: {
+        width: "50%",
+    },
+    buttonWrapper: {
+        width: "100%",
+        display: "flex",
+        justifyContent: "center",
+        marginTop: "3%",
+        padding: "0% 43%",
+        height: "40px",
+        color: "primary",
+    },
+    button: {
+        width: "100%",
+    },
+}));
