+++ src/Components/News/SearchSub/Input/InputStyles.tsx
@@ -1,0 +1,13 @@
+import { createStyles } from "@mantine/core";
+
+export const useStyles = createStyles(() => ({
+    input: {
+        width: "50%",
+    },
+    chips: {
+        width: "100%",
+        display: "flex",
+        justifyContent: "space-between",
+        padding: "0% 20%",
+    },
+}));
